feat: use informer to watch and cache connectors
As today, the synchronizers determines the highest resource revision to
be used to poll the fleet manager by loading all the ManagedConenctor
CRs from the API server and then extracting the highest resource
revision. This change is aimed at leveragin the Informer to cache the
ManagedConenctor CRs in memory and reduce the interaction with the API
server.