MGDCTRS-78: migrates to java operator sdk v2
This migrates to the new Java Operator SDK, version 2.0.2, and also the
Quarkus Operator SDK, to version 3.0.2. For the controller and its
reconcile methods, it's mostly name changes. For the EventSources there
has been a moderate change to the API. For e.g., Events are now
binded to a single ResourceID (before a resource selector). The new
API suggests the use of Informers for EventSources, which would make
for a even simpler implementation, but Informers have internal cache
which could represent a high memory usage in scenarios with many
connectors deployed, that's why we are still using our own Watch-based
EventSources. Existing EventSources had to be changed to trigger
multiple events because of the new 1-1 resource-event relation.