fix(MGDCTRS-458): the reify process fails to generate a valid kamelet binding
The root cause of the issue is in the wrong usage of KubernetesResource
as such interfce should only be used for real resources. In our camel
model, some of the objects that are part of the KameletBinding resource
are also marked as KubernetesResource and that instruct Jackson to use a
special deserializer.

As for the fabric8 kubernetes client 5.12.x, a new defensive copy of the
reources has been introduced for operations like create or update and
since such copy is perfomed by serializing and then deserializing the
resource using jackson, the deserializer configured by the
KubernetesResource kicked in and cause the resource to be wroingly
serialized.