feat(MGDCTRS-639): makes operators cluster-wide
Changes the behavior of debezium and camel operators so they will watch
the entire cluster for changes on ManagedConnector and other resources.

The property cos.connectors.namespace is not used anymore. Resources are
now created in the same namespace where the ManagedConnector resource
already is.

Roles and RoleBindings were changed to cluster.