MGDCTRS-974: Set limit ranges assigned to a kubernetes namespace
With this PR a new `LimitRange` resource is created in any namespace
where resource quota are configured assuming that at this stage,
only eval namespace have such quota set.

The reason of such resource to be created is that,  when allocating
compute resources, each container may specify a request and a limit
value for either CPU or memory and the quota can be configured to
quota either value per namespace but if the quota has a value specified
for requests.cpu or requests.memory, then it requires that every
container makes an explicit request for those resources.

If the quota has a value specified for limits.cpu or limits.memory,
then it requires that every incoming container specifies an explicit
limit for those resources.

Since not our connector do explicit specify limits, a `LimitRange` can
be used to automatically set them:

```yaml
apiVersion: v1
kind: LimitRange
metadata:
  name: redhat-openshift-connectors-ca5j6vufrusf747mmmmg-ranges
spec:
  limits:
  - default:
      cpu: 500m
      memory: 500m
    defaultRequest:
      cpu: 200m
      memory: 128m
    type: Container
```

With this set-up each new container that is created on the eval namespace,
will receive a default limit computed taking into account the limits
sets on the namespace and the number of connectors that are allowed to
be created to the namespace.